feat(ui5-message-strip): introduce custom colors (#8757)
With the following enhancement, the end user will be able to create message strips with custom colors from a predefined set of colors. That is achieved enhancing the `design` property enum `MessageStripDesign` with two new options `ColorSet1` and `ColorSet2` as well as the newly introduced property `colorScheme`.
Setting the `design` property value to ColorSet1` or `ColorSet2`, then you are able to choose between 10 different colors for each using `colorScheme` property with values from "1" to "10".