fix(ui5-*): Make setAnimationMode() properly change modes
Compare the value of the passed in animationMode to the values of the
AnimationMode enum (lowercase) rather than to the property (key) names
(uppercase). This will make it so that explicitly passing in a mode
(e.g. AnimationMode.None) or passing in the string value corresponding
to a mode (e.g. "none") will correctly result in the current mode being
set to the lowercase string value, which will make it possible to use
setAnimationMode() to disable component animations as intended.

Fixes #8964