chore: allow setting string to Popover#opener in TS env (#9004)
In TS env currently one can't set string id as opener declaratively as the TS type definition is restricted to HTMLElement.
With this change we include "string" into the accepted parameters for the "opener" property. Some of the code had to be adjusted as previously used to work with HTMLElement only.