feat(ui5-menu-item): add endContent slot
New slot `endContent` is introduced in `ui5-menu-item` component. This slot accepts components that are displayed at the end of the Menu Item if it does not have sub-menu. If there are components slotted and also `additionalText` is being set, the `additionalText` will not be shown as well. 

It is strongly recommended  to slot only icon-only `ui5-button`, `ui5-icon` or `ui5-link` components in order to preserve the intended design.

BGSOFUIBALKAN-8160
#6350