fix(ui5-special-date): add visual representation for Working/NonWorking types (#9524)
Previously, when adding a `SpecialCalendarDate` to the `Calendar` component and specifying the public types `Working` and `NonWorking`, these settings were not properly applied.

For example, setting a `Working` type to a weekend day did not visually reflect that the day was designated as a working day and vice versa for `NonWorking` type.

With this change we resolve that issue, ensuring that the specified types (`Working` and `NonWorking`) now take effect as intended.