feat(ui5-ai-button): initial implementation
With this PR we introduce a new type of a Button suitable for AI-related senarios.

There are 2 components - `ui5-ai-button` and `ui5-ai-button-state` that are in EXPERIMENTAL state.

The Button can have one or more states predefined.

The transition between the states can happen on click of the Button and there are fancy animations of icons, texts and button size on state change.

Each state can have the following properties set:

**text** - text of the button
**icon** - the icon before the text
**endIcon** - the icon after the text

There is **click** event fired when the Button is being clicked. This event can used from the developer to set different state of the button and to execute some state-specific functionality.