refactor(ui5-special-date): remove calendar date inheritance (#9237)
Removing the connection between special dates and calendar dates, because the calendar date now implements an interface specific to date selection in the calendar. The current implementation will show the special dates as a type that's feasable for Calendar date selection and after this change, the selected date will no longer show as component implementing the ICalendarSelectedDates interface.