chore(ui5-avatar): remove _size proeprty (#9362)
Ever since the deprecation of the avatar's group avatarSize property (https://github.com/SAP/ui5-webcomponents/pull/3229) the _size property of the avatar is not used. Since recent changes those underscored properties are showing as the attributes of the component. This causes the avatar size to be displayed incorrectly duo to the default value of the _size proeprty.