fix(ui5-table): correct load-more event on mode change (#9434)
If the `ui5-table` is made with `mode=Growing`, and then the mode is changed to `None` and back to `Growing`, the `load-more` event is not fired anymore. This is because the observed element by IntersectionObserver is not in the DOM anymore because it is conditionally rendered. To fix this, we reset the `tableEndObserved` flag when the mode changes, so the observed element is set again correctly when the DOM is ready.

Fixes: https://github.com/SAP/ui5-webcomponents/issues/8052