fix(ui5-select): infinite attribute update in value property setter (#9500)
An infinite attribute update occurred due to the set and get accessors of the value property when an invalid value was passed (one that did not match any of the select's options). This caused the getter to return a value out of sync with what was set by the setter (the attribute that results in the DOM by the set) prompting the framework to trigger another set.

This PR remove the value attribute from the DOM. This prevents the framework on attribute check call and the infinite recursive call stops.