fix(ui5-dynamic-page): fixed wrong overlapping in header area
There was a problem where the handle of the Slider components was
overlapping over the Dynamic Page header area. This was due to equal
z-index of both the handle and the Dynamic Page header wrapper.

This change increases the z-index of the header wrapper making sure that
the handle of the slider is always rendered under it and not overlapping.

Fixes: #9560