test: fail on dev errors (#9519)
When `assert` was failing it was preventing the `test.callback` execution and some development logs were ignored. `test.callback` is the test body and it is marked as done once the test is executed. Now, when `[UI5-FWK]` logs are detected call the callback again which leads to an error that says the test is tried to be finished multiple times.