feat(ui5-li-notification-group): add new features - growing and sticky header
According to the new design the notification groups should be able to
grow by pressing "More" button.
Then the "load-more" event will be fired and more notifications
can be loaded.
The other new feature is the sticky header, which stays
on top of the popover with notifications,
when the user scrolls down. When next group is reached the sticky header
will be replaced with the
new one corresponding to the currently visible group.

JIRA: BGSOFUIRODOPI-3224