fix(ui5-menu): fix focus steal of ui5-wizard in mozilla (#9845)
Previously in Mozilla Firefox browser the ui5-menu was 'stealing' the focus of the ui5-wizard resulting in scroll back through the steps.
The problem occured when the ui5-wizard contained a menu, with or without ui5-menu-item, and sub-menus. The reason was due to some timing issue, when executing the _createSubMenu() --then--> _openItemSubMenu() methods in the ui5-menu.