fix(ui5-list): growing button loading layout fixed
With #9977 change we started rendering the busy indicator next to the
growing text.

This introduced an issue where we align the growing text slightly to the
right of the busy indicator but before the busy indicator is actually
shown.

This fix fixes the issue and the growing text is centered before
showing the busy indicator.