feat(ui5-barcode-scanner-dialog): added support for custom header and footer slots
This update introduces two new slot attributes for
`ui5-barcode-scanner-dialog`, allowing developers to customize the header
and footer of the dialog. If no custom content is provided, the dialog
will use its default layout. The new slots enhance flexibility for
developers, enabling the addition of custom titles, buttons, and any
other elements in the header and footer areas.

Related: #8919