docs: turn off forced colors mode in website (#10085)
Windows Contrast themes adjust colors of UI5 Web Components which leads to poor visual. 
The PR fixes the issue for the playground samples setting `forced-color-adjust` to `none`. The property allows authors to opt certain elements out of forced colors mode as explained in [MDN](https://developer.mozilla.org/en-US/docs/Web/CSS/forced-color-adjust).
And adds suggestion for fixing it in apps the same way. We don't own the page and we believe this is in the apps domain to set it.

Fixes: https://github.com/SAP/ui5-webcomponents/issues/9436 https://github.com/SAP/ui5-webcomponents/issues/9917