fix(ui5-table-growing): button is shown despite scroll type (#10142)
When the table is not filling/utilizing all the space inside of the scroll container, we need to show a "More" button, as scrolling is not possible yet.
Once the scroll container is scrollable, the "More" button should disappear, which does not happen as of now.
Adjusted the logic determining when to render a button. Additionally, removed the overflow on the host element. Instead the table element should fill the parent element now.

Fixes #10045