fix(ui5-dialog): width on mobile is not bigger than the phone width (#10199)
Issue: When the base font in the 'html' element is changed to something
different than the default 16px (for example 32px) the width
of the dialog can become larger than the width of the phone's
display. This is caused by the 'min-width' (20rem) of the dialog.
For the dialog on phone it is recommended by the design to set the dialog's
property 'stretch' to true to use the full screen size.

The solution: When we have stretched dialog on phone the 'min-width' should
 not be applied (the width is 100%).

fixes: #10000
downport of #10199