fix(build): generation of custom-elements.json (#10403)
The API generation was failing because event types are retrieved from a class member named `eventDetails`. To identify this, we needed to iterate through all class members and check their names. However, the constructor is also part of class members, but it is not treated as member by typescript because it doesn't have name.