fix(ui5-popover): allow setting `opener` to an abstract element DOM reference (#10419)
When Popover's opener is set as a DOM reference and if happens to be a UI5 Element, we need to return its DOM (focus) reference. As of now we just return the element itself. While this works for most of the cases, it does not work for abstract elements as the popover won't be able to calculate its position due to not having a physically DOM element as opener.

The case when the opener is set as string, ID, pointing to an element has been already handled. For this case we already call getFocusDomRef. And, this change addresses the other case - setting the opener as DOM reference.