fix(ui5-combobox): fire selection-change event when typed word fully matches item
Fixes #10239

When combobox is in noTypeahead mode, if the user writes a text that matches exactly
a list item, selection-change event was not fired if the user selects the item by
clicking on it.
Now the selection0change event fires when item text is exactly typed and
immediately selected from the items list.