fix(ui5-split-button): fix doubled click event when space is pressed (#10241)
When a default action of a ui5-split-button was activated by pressing of Space key, the click event was fired twice - on keydown and on keyup. It should be fired only once on keyup only if the Space is pressed and before the release Escape or Shift is not pressed (this is accessibility feature related to all buttons).

This PR fixes the issue.