chore: correct generation of events in cem (#10827)
TypeScript decorators can be imported in two ways:  
- **Default import** from `@ui5/webcomponents-base/dist/decorators/{decoratorName}.js`  
- **Named import** from `@ui5/webcomponents-base/dist/decorators.js`  

Recently, the `event.js` decorator was deprecated and replaced with `event-strict.js`. When using the default import, the decorator could retain its old name, and the custom element analyzer would still process it correctly. However, when using the named import, the `@eventStrict` decorator was not processed, leading to missing parts of the component's API.  

This PR adjusts the custom element analyzer to correctly process both `@event` and `@eventStrict` decorators.