refactor: vite.config.js for enhanced flexibility and tool compatibility (#10818)
Rework the site configuration to avoid returning an asynchronous function, making it easier to override when needed.  

In addition, add a new `vite.config.js` file to each newly created package. This allows other tools, such as Cypress, to detect Vite’s configurations, which was not possible when using them directly from the `@ui5/webcomponents` package.