fix(ui5-switch): remove delegate focus of switch (#10798)
With change #10588 we addressed an issue, where the focus was being applied even if the root of the element (the touch area) was not clicked/touched, by removing the pointer events on host level.

However this would've caused issues within cypress testing.

To address the issue without breaking the cypress tests, we now remove the `delegateFocus` option, from the `shadowRootOptions`. By removing focus delegation, we stopped the automatic forwarding of focus. Now, only an explicit click on the focusable inner element results in it receiving focus.