fix(ui5-shellbar): fix separators visibility when items are hidden (#11058)
When items are being hidden, the disappearing separators were messing up the math for calculating available space.

This commit fixes the issue by rendering the hidden separators with the last visible item. This way the space is still resevered and taken into account when calculating the available space.

The tests are also extended to cover this case.