fix(ui5-menu): fix menu item type fix (#11129)
Before:
When used in SideNavigation, the item doesn't have "ui5-menu-item" attribute, because it's a subclass of MenuItem and the navigation checks are failing.

After:
We check with a getter done in the MenuItem class, so even if the class is extended, we will have the correct results in these checks.

fixes: #10933