fix(ui5-menu): handle uncaught error on mouseout (#11122)
Previosuly when the menu opener was not being the correct type/item there was an error being thrown in the console.

This was happening because in some cases (e.g menu items with subitems used in `ui5-side-navigation`) the opener of a submenu happens to be an `ui5-icon` or `<a>` tag for example. Therefore in those cases the items lacked properties (such as `associatedItem` as we can see from the screenshot below) leading to `item = undefined`

We fix the problem, by assigning the correct target upon mouse out, which is handled in the `_getSubmenuReference()` method.