fix(ui5-calendar): remove auto focus of current day on initial render (#11213)
Previously when initially rendered the `ui5-calendar` component was "stealing" the focus of the page as well as when being re-rendered since the focus was being applied in the `onAfterRendering()` hook.

With this change we now handle this behaviour, by applying focus only when the `ui5-calendar` is intentionally focused by the end user, ensuring that the focus is not being "stolen" from other interactive web page elements while still ensuring that the correct day is focused.

Fixes: #10728