chore: enhance cypress test configuration with conditional execution of tools (#11195)
Previously, the Cypress code coverage plugin was used even when a coverage report was not required. Additionally, Cypress accessibility testing (introduced in https://github.com/SAP/ui5-webcomponents/pull/11128) is intended to run only when a report is needed and requires an environment variable. This PR introduces new internal flags in the `package-script.cjs` file of each component package, simplifying the process of setting the required environment variables when these features need to be enabled.