fix(framework): fix fetching "id" (indonesian) language (#11251)
If someone sets "id" (indonesian) as a language, he gets the english texts.
We do have messagebundle_id.properties and regularly update it. However we did not emit the messagebundle_id.json that should have been fetched over the network, because we missed to describe it as one of the supported languages. There we described the old code "in", which will work. But, at some point we stopped receiving translations for the in code (we did not notice that, because we receive a huge PR including 100 of files and "_in.properties' was one of them) for 4 years, which means it's not supported any more and we can safely remove the messagebundle_in.properties file ad it's also well behind the messagebundle_id.properties which is updated just recently.

Fixes: #11233