chore: stable failing test (#11303)
### UploadCollectionItem  
When the detail button is pressed, `editing` mode is activated. In this state, the element with the `.ui5-uci-file-name` class is not rendered in the DOM.

### Search (Mobile)  
Ensure that the responsive popover is open before interacting with any components inside it.

Skip test due bug inside component https://github.com/SAP/ui5-webcomponents/issues/11314

### Responsive Popover Command  
The responsive popover command should locate a `ui5-popover` or `ui5-dialog` within the shadow root and verify that it is internally visible.