chore: rename tsconfig in new package script (#11322)
The `vite.config.js` file collects and processes all `tsconfig.json` files. During this process, it also detects newly added `tsconfig.json` files inside the `create-package/template` folder, which leads to errors being shown in the console when the development server starts. To address this, the files are now renamed to `tsconfig.template.json` and are reverted back to `tsconfig.json` when the generation script is executed.