feat(ui5-menu): introduce HorizontalAlign property (#11313)
This change introduces a new `horizontalAlign` property to `ui5-menu`, similar to what's available in `ui5-popover`. It allows users to control the menu's horizontal alignment.

The default value is `start`, but it can also be set to `end`, `center`, or `stretch`.

Usage:
Setting the property to end as follows will align the menu to the end of the opener button:

```HTML
<ui5-menu horizontal-align="End">
    <ui5-menu-item ...
</ui5-menu>
```

Fixes: #10765