feat(ui5-shellbar): introduce async search btn getter (#11338)
The `searchButtonDomRef` has been refactored into an asynchronous method `getSearchButtonDomRef` that waits for the component to finish rendering. This ensures the DOM is available for use cases that require focusing the search button.

A similar code has been applied to the `search-field-toggle` event, which is now fired after rendering is complete.