fix(ui5-breadcrumbs): dropdown arrow focused color fixed (#11362)
Bug description: When we have dropdown arrow (when there is not enough
space for all the links) and we focus it, the color of the arrow and
triple dots remains bluish. It should become white.

This change fixes the issue by improving the CSS selector for making
the focused color to be bright.