fix(ui5-side-navigation): add parent item to submenu
According to the interaction design clicking on the parent item in the Overflow menu
should only open the sub-menu, but should not select the parent item.
Clicking on the parent item in the sub-menu selects it and closes the menu -
this is the only way to select the parent item in the overflow menu
(the menu lacks two-click area functionality for the moment).
This is the reason why we add again the parent item also in the sub-menu.

JIRA: BGSOFUIRODOPI-3443