fix: escape backslash when processing component's css (#11425)
The `.w-\[100px\]` selector is a valid CSS class. However, when processed by the tools package, it is incorrectly saved as `.w-[100px]`, making it an invalid selector.  
With this PR, when a backslash is encountered during processing inside a CSS file, it is properly escaped before being saved to the new file.

Fixes: https://github.com/SAP/ui5-webcomponents/issues/11413