fix(ui5-calendar-legend): align selected item representation to new VD spec (#11391)
We have new design specification, about the representation of the Selected day default item of the ui5-calendar-legend component.

Now in SAP Horizon and SAP Horizon Dark themes, the Selected Day item of the ui5-calendar-legend has a dot in the middle, which makes its concept more clear.

Before
2025-04-23_13-43-56

After
2025-04-23_13-49-25