fix(compat/ui5-table): fire loadMore with arrow-down on last row (#11401)
On some screen sizes and table row's heights, the user can't reach the bottom of the table by using the arrow-down key.
Although the last row is being reached and gets focused, the very bottom of the table can't be reached, and the load-more event isn't fired.
For this use-case, but not only, we now fire the "load-more" event whenever the user presses ARROW DOWN on the last table row.

Fixes: #9235