fix(ui5-button): announce accessible name (#11396)
Issue:
- The badge wasn't announced at all when an accessible-name attribute is provided
for the ui5-button. The reason behind this behaviour is that the screen reader doesn't
read the inner text nodes of the root button node when the same has aria-label attribute.

Solution:
- The proper badge description text is placed into the aria-label attribute of the root
button element. As a result the screen reader won't rely on the inner text content
for the announcements.
- Additionally the design description is also moved into the aria-label attribute,
in order to keep a proper announcement order.

Fixes: #11350