chore: enable hydration for declarative shadow DOM components (#11488)
This PR enables hydration of declarative Shadow DOM.

When the server responds with a custom element containing a `<template>` tag and it is connected to the DOM, the element is instantiated with a Shadow DOM.

To avoid a full re-render, the DOM sent from the server is preserved during the initial render and hydrated to enable interactivity.

Since the server must return CSS styles within a `<style>` element (or `<link>`, though currently only `<style>` is supported) to ensure correct visual appearance, these styles are removed after hydration to prevent duplication. The component will add its styles to `adoptedStyleSheets` once it is added to the DOM.

Part of: https://github.com/SAP/ui5-webcomponents/pull/11488