fix(ui5-list): adjust observer to handle sticky headers (#11526)
Problem:
The list's onLoadMore event is not triggered in Firefox at certain zoom levels when a sticky header is present. This is due to how Firefox calculates intersections with sticky positioned elements, causing subpixel rendering issues.

Solution:
Use the proper scroll container as the root for the IntersectionObserver and add a rootMargin of 5px.

Similar to: #11242
Fixes: #11461