fix(ui5-color-palette): show tooltip on focused hovered items (#11519)
There was a bug where upon hovering on a focused item, the tooltip (if such is present) was not showing.
The reason behind that was a CSS property disabling all events - pointer-events: none.

We fix that by ensuring that the property pointer-events: none is present when the ui5-color-palette-item (swatch) is focused but NOT hovered. This ensures that:

 - The tooltip is visible when the user hovers over a focused ui5-color-palette-item;
 - The focus/hover visual appearance is maintained;