fix(ui5-split-button): fire click event when focused with shift + tab (#11197)
### Main Problem
There was an issue, where if the `ui5-split-button` was to be focused by using `Shift + Tab` combination, the `click` event wasn't firing, after the first time a default action was executed (pressed `Enter` or `Space`).

The reason behind this, was a flag called `_shiftOrEscapePressed`.
Since the component has a handler for `focus-in` event which was just early returning if the component `matches` the `:focus-within` CSS selector, which was causing the flag to NOT reset. Therefore the component wasn't firing `click` event, due to the condition:

```ts
_fireClick(e?: Event) {
   e?.stopPropagation();
   if (!this._shiftOrEscapePressed) {
      this.fireDecoratorEvent("click");
   }
   this._shiftOrEscapePressed = false;
}
```

### Solutions

 - We solved few issues, by introducing a new flag called `_shiftOrEscapePressedDuringSpace`, which handles the cases only when `Shift` or `Escape` keys are pressed simultaneously with the `Space` Bar.
 - Deleted an unused forgotten flag;
 - Handled an issue where the arrow button was being 'frozen' if simultaneously pressing `ArrowUp/Down` keys and `Tab/Shift + Tab`.
 - `Shift/Escape` keys now correctly prevent events from firing;
 - Fixed a bug where `arrow-button` fired `click` **AND** `arrow-click` events.