feat: add `Assets-node` module for node/ssr env (#11522)
**Change**

The change provides new entry point for assets (cldr, i18n and theming), called `Assets-node.js`.
This module is an alternative to `Assets.js`- the assets are only registered and loaded on the fly with dymamic imports when needed **with added support for the** `with: { type: 'json' }` import attribute, which is required in certain environments, such as Node.js with server-side rendering (SSR) - to properly load JSON files.

**Example**:
```ts
await import("../assets/i18n/messagebundle_bg.json", { with: { type: 'json' } });
```

**Recommended usage**: in environments that require SSR compatibility (e.g., Node.js).

**Known issues**
- Webpack < 5.84: older versions of Webpack do not support the `with: { type: 'json' }` syntax and will fail to build.
- Vite (dev mode): When using Vite, the development server may raise errors (e.g., expecting application/json) unless additional configuration is added to handle JSON imports with attributes.