fix(ui5-button): prevents native behavior when click prevented (#11678)
Native click behavior triggered by the `ui5-button` is not suppressed based on the custom event fired by the web component. Although the native `click` event does not propagate, its default behavior still occurs—even when the custom event is prevented.
With this PR, the native behavior is now correctly prevented when the custom event is prevented.

Fixes: https://github.com/SAP/ui5-webcomponents/issues/11653